requirements.txt
transformers
einops
pillow
torch
torchvision
numpy<2.0
https://github.com/KurimuzonAkuma/pyrogram/archive/dev.zip
TgCrypto
python-dotenv
